Roll SqlConfigurableRetryLogicManager.NetCoreApp.cs into SqlConfigurableRetryLogicLoader class
