+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Reliability/SqlConfigurableRetryLogicLoader.cs
@@ -8,6 +8,11 @@ using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using System.Text.RegularExpressions;
 
+#if NET
+using System.IO;
+using System.Runtime.Loader;
+#endif
+
 namespace Microsoft.Data.SqlClient
 {
     /// <summary>
@@ -38,18 +43,20 @@ namespace Microsoft.Data.SqlClient
         /// </summary>
         internal SqlRetryLogicBaseProvider CommandProvider { get; private set; }
 
-        public SqlConfigurableRetryLogicLoader(ISqlConfigurableRetryConnectionSection connectionRetryConfigs,
-                                               ISqlConfigurableRetryCommandSection commandRetryConfigs,
-                                               string cnnSectionName = SqlConfigurableRetryConnectionSection.Name,
-                                               string cmdSectionName = SqlConfigurableRetryCommandSection.Name)
+        public SqlConfigurableRetryLogicLoader(
+            ISqlConfigurableRetryConnectionSection connectionRetryConfigs,
+            ISqlConfigurableRetryCommandSection commandRetryConfigs,
+            string cnnSectionName = SqlConfigurableRetryConnectionSection.Name,
+            string cmdSectionName = SqlConfigurableRetryCommandSection.Name)
         {
-#if NET
+            #if NET
             // Just only one subscription to this event is required.
             // This class isn't supposed to be called more than one time;
             // SqlConfigurableRetryLogicManager manages a single instance of this class.
             System.Runtime.Loader.AssemblyLoadContext.Default.Resolving -= Default_Resolving;
             System.Runtime.Loader.AssemblyLoadContext.Default.Resolving += Default_Resolving;
-#endif
+            #endif
+            
             AssignProviders(connectionRetryConfigs == null ? null : CreateRetryLogicProvider(cnnSectionName, connectionRetryConfigs),
                             commandRetryConfigs == null ? null : CreateRetryLogicProvider(cmdSectionName, commandRetryConfigs));
         }
@@ -156,10 +163,12 @@ namespace Microsoft.Data.SqlClient
         }
 
         private static object CreateInstance(
-#if NET
+            #if NET
             [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicParameterlessConstructor | DynamicallyAccessedMemberTypes.PublicMethods)]
-#endif
-            Type type, string retryMethodName, SqlRetryLogicOption option)
+            #endif
+            Type type,
+            string retryMethodName,
+            SqlRetryLogicOption option)
         {
             string methodName = nameof(CreateInstance);
             SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Entry point.", TypeName, methodName);
@@ -324,5 +333,103 @@ namespace Microsoft.Data.SqlClient
             }
             return new HashSet<int>();
         }
+        
+        #if NET
+        #region Type Resolution
+        
+        private static Assembly AssemblyResolver(AssemblyName arg)
+        {
+            string methodName = nameof(AssemblyResolver);
+
+            string fullPath = MakeFullPath(Environment.CurrentDirectory, arg.Name);
+            SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Looking for '{2}' assembly by '{3}' full path."
+                , TypeName, methodName, arg, fullPath);
+
+            return fullPath == null ? null : AssemblyLoadContext.Default.LoadFromAssemblyPath(fullPath);
+        }
+        
+        /// <summary>
+        /// Load assemblies on request.
+        /// </summary>
+        private static Assembly Default_Resolving(AssemblyLoadContext arg1, AssemblyName arg2)
+        {
+            string methodName = nameof(Default_Resolving);
+
+            string target = MakeFullPath(Environment.CurrentDirectory, arg2.Name);
+            SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Looking for '{2}' assembly that is requested by '{3}' ALC from '{4}' path."
+                , TypeName, methodName, arg2, arg1, target);
+
+            return target == null ? null : arg1.LoadFromAssemblyPath(target);
+        }
+        
+        /// <summary>
+        /// Performs a case-sensitive search to resolve the specified type name 
+        /// and its related assemblies in default assembly load context if they aren't loaded yet.
+        /// </summary>
+        /// <returns>Resolved type if it could resolve the type; otherwise, the `SqlConfigurableRetryFactory` type.</returns>
+        private static Type LoadType(string fullyQualifiedName)
+        {
+            string methodName = nameof(LoadType);
+            SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Entry point.", TypeName, methodName);
+
+            var result = Type.GetType(fullyQualifiedName, AssemblyResolver, TypeResolver);
+            if (result != null)
+            {
+                SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> The '{2}' type is resolved.",
+                                                        TypeName, methodName, result.FullName);
+            }
+            else
+            {
+                result = typeof(SqlConfigurableRetryFactory);
+                SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Couldn't resolve the requested type by '{2}'; The internal `{3}` type is returned.",
+                                                        TypeName, methodName, fullyQualifiedName, result.FullName);
+            }
+            SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Exit point.", TypeName, methodName);
+            return result;
+        }
+
+        /// <summary>
+        /// If the caller does not have sufficient permissions to read the specified file, 
+        /// no exception is thrown and the method returns null regardless of the existence of path.
+        /// </summary>
+        private static string MakeFullPath(string directory, string assemblyName, string extension = ".dll")
+        {
+            string methodName = nameof(MakeFullPath);
+            SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Looking for '{2}' assembly in '{3}' directory."
+                                                    , TypeName, methodName, assemblyName, directory);
+            string fullPath = Path.Combine(directory, assemblyName);
+            fullPath = string.IsNullOrEmpty(Path.GetExtension(fullPath)) ? fullPath + extension : fullPath;
+            SqlClientEventSource.Log.TryTraceEvent("<sc.{0}.{1}|INFO> Looking for '{2}' assembly by '{3}' full path."
+                                                    , TypeName, methodName, assemblyName, fullPath);
+            return File.Exists(fullPath) ? fullPath : null;
+        }
+
+        private static Type TypeResolver(Assembly arg1, string arg2, bool arg3)
+        {
+            IEnumerable<Type> types = arg1?.ExportedTypes;
+            Type result = null;
+            if (types != null)
+            {
+                foreach (Type type in types)
+                {
+                    if (type.FullName == arg2)
+                    {
+                        if (result != null)
+                        {
+                            throw new InvalidOperationException("Sequence contains more than one matching element");
+                        }
+                        result = type;
+                    }
+                }
+            }
+            if (result == null)
+            {
+                throw new InvalidOperationException("Sequence contains no matching element");
+            }
+            return result;
+        }
+        
+        #endregion
+        #endif
     }
 }
